Use organization_list with all_fields=True instead of organization_show
CKAN's organization_list() (when called with the all_fields=True param)
and organization_show() currently return organization dicts in different
formats. Specifically, in the dicts from organization_list() "packages"
is the number of packages, whereas in the dicts from organization_show()
"packages" is a list of package dicts.

CKAN core uses organization_list() with all_fields=True (eg. when
showing organizations on the organization index page), so change this
extension to use that as well, instead of using organization_show().

This means that it's no longer necessary to override the core
organization_item.html template to get the organizations to render
properly on the front page.

This also fixes a problem with an organization's number of packages not
rendering properly on the /organization page on DataGM.